organize code and travis to remove scipy dependence
