+++ .travis.yml
@@ -1,12 +1,15 @@
 sudo: false
+notifications:
+  email: false
 language: python
 python:
   - "2.7"
   - "3.4"
   - "3.5"
   - "3.6"
-notifications:
-  email: false
+env:
+  - DEPS="pip nose future numpy scipy"
+  - DEPS="pip nose future numpy"
 before_install:
   - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
       wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
@@ -18,6 +21,6 @@ before_install:
   - conda update --yes conda
   - conda config --add channels conda-forge
 install:
-  - conda install --yes python=$TRAVIS_PYTHON_VERSION pip numpy scipy nose future
+  - conda install --yes python=$TRAVIS_PYTHON_VERSION $DEPS
   - pip install -v .
 script: nosetests tests
